Make pulp->avg_size an array
Intended for there to be one per vertex weight. If g is unweighted, it
will just allocate a size 1 array and use g->n like normal.

There are some possibly bad pieces of code here; the only reason that
avg_size needed to become an array is because vertex_weights_sum needed
to become an array. As such, a lot of places that previously assumed
avg_size was a single number now just use avg_size[0].